Clean up app code.
Remove old and unused App messages and video player output.

Removed frame info from app since it is never used.

Use single variable for showing spinner and video. Easier to track logic of when it is set.

Reset preview. May not be fully resetting right now.

Player has flag for seeing if pipeline ready. Is set when given a uri and unset when pipeline reset.