Don't save thumbnails to disk and instead pass them as texture to populate timeline.
Saves on cost of saving to disk and leaving disk footprint.

Initially wanted it to save time of down sampling as well, but with native res frames ui is stuttering.

Might not be saved in order right now. Looking into making the down sampling a lot faster (rn ~900ms for 4k frame) and then might remove the thread spawning.