Add image adjustments (contrast) and previewing of adjustment.
Have a writeable buffer and buffer that is used in shaders. Writeable is copied into shader buffer. Size are constant with effects supported.

Should copy only when some change in parameters, and only then should the writeable buffer be updated.

Default values should produce no change in output, as there is no flag for detecting if effect active. For that would need either separate buffer to flag or a uniform buffer, but then each effect has to have the same max of parameters.

Currently only contrast hooked up. UI is just for testing.