Implement fill mode for slider.
Edge to Edge is and previously implemented. Left hand side is treated as 0% and right hand as 100%.

Center out is treating the 50% mark the midpoint. Values are filled in from 50% towards the value.
For example 45% will have 5% of the slider from 50% to 45% filled in, and likewise 65% will have 15% of the slider from 50 to 65.

Center out will likely become the default as most sliders will have a neutral at 0 and ability to increase and decrease the effect.

Make show ticks the default