Orient videos that have orientation set in metadata image-direction tag correctly.
Thumbnails are presented in wrong aspect ratio currently.

Some videos are recorded as 16:9 and store metadata to signify they are actually 9:16.

Now look for the metadata tag and store the orientation in metadata collection. Store it as a base angle in orientation struct. New method absolute angle to get sum of applied angle and base angle.

Change should not affect anything which has a base angle of 0. So any bugs would affect videos that were already oriented incorrectly.

Rename is_vertical to is_width_flipped to be more clear in what function is checking. Would be confusing for videos that are in a vertical aspect ratio natively.

Do not know all possible permutations of image-orientation tag so parsing it into my Orientation struct is incomplete.